refactor: change MaxFramesInFlight and CurrentFrame to uint

+++ src/PanguEngine/Rendering/Vulkan/VulkanBuffer.cs
@@ -50,7 +50,7 @@ public sealed unsafe class VulkanBuffer
 
         var buffer = Buffer;
         var allocation = _allocation;
-        var retireValue = VulkanContext.GlobalTimelineValue + (ulong)VulkanContext.MaxFramesInFlight;
+        var retireValue = VulkanContext.GlobalTimelineValue + VulkanContext.MaxFramesInFlight;
         VulkanDeletionQueue.Enqueue(retireValue, () => VulkanAllocator.DestroyBuffer(buffer, allocation));
     }
 }
